Move condition related store files into separate file + add tests
