+++ src/lib/state/scene_data.spec.ts
@@ -8,7 +8,6 @@ import {
 	setCombatantCombatStateToDead,
 	setCombatantCombatStateToExpecting,
 	setCombatantCombatStateToWaiting,
-	toggleCondition,
 	type Combatant,
 	type Scene
 } from './scene_data.svelte';
@@ -202,104 +201,6 @@ describe('moveCombatantToTick', () => {
 			expect(combatant).toEqual(lastActiveCombatant);
 		}
 	);
-});
-
-describe('toggleCondition', () => {
-	it('toggle condition', () => {
-		givenScene(TEST_SCENE_SMALL);
-
-		// Add bleeding I
-		toggleCondition('MerryId', 'bleeding');
-		expect(sceneData.combatants.find((c) => c.id === 'MerryId')?.conditionStates).toEqual([
-			{
-				activeLevel: 1,
-				id: 'bleeding',
-				activeSinceTick: 3
-			}
-		]);
-
-		// Change to bleeding level III
-		toggleCondition('MerryId', 'bleeding');
-		toggleCondition('MerryId', 'bleeding');
-		expect(sceneData.combatants.find((c) => c.id === 'MerryId')?.conditionStates).toEqual([
-			{
-				activeLevel: 3,
-				id: 'bleeding',
-				activeSinceTick: 3
-			}
-		]);
-
-		// Toggle one more time to remove bleeding
-		toggleCondition('MerryId', 'bleeding');
-		expect(sceneData.combatants.find((c) => c.id === 'MerryId')?.conditionStates).toEqual([]);
-	});
-
-	it('toggle condition and continue ticks', () => {
-		givenScene(TEST_SCENE_SMALL);
-
-		// Add bleeding II for Merry
-		toggleCondition('MerryId', 'bleeding');
-		toggleCondition('MerryId', 'bleeding');
-		expect(sceneData.combatants.find((c) => c.id === 'MerryId')?.conditionStates).toEqual([
-			{
-				activeLevel: 2,
-				id: 'bleeding',
-				activeSinceTick: 3
-			}
-		]);
-		expect(sceneData.combatants.find((c) => c.id === 'GandalfId')?.conditionStates).toEqual([]);
-
-		// Verify that Gandals condition is unaffected by Merry's tick progression
-		moveCombatantByTicks('MerryId', 5);
-		toggleCondition('GandalfId', 'burning');
-		expect(sceneData.combatants.find((c) => c.id === 'MerryId')?.conditionStates).toEqual([
-			{
-				activeLevel: 2,
-				id: 'bleeding',
-				activeSinceTick: 3
-			}
-		]);
-		expect(sceneData.combatants.find((c) => c.id === 'GandalfId')?.conditionStates).toEqual([
-			{
-				activeLevel: 1,
-				id: 'burning',
-				activeSinceTick: 4
-			}
-		]);
-
-		// Advance Gandalf and verify both combatants
-		moveCombatantByTicks('GandalfId', 5);
-		toggleCondition('GandalfId', 'burning');
-		toggleCondition('GandalfId', 'burning');
-		expect(sceneData.combatants.find((c) => c.id === 'MerryId')?.conditionStates).toEqual([
-			{
-				activeLevel: 2,
-				id: 'bleeding',
-				activeSinceTick: 3
-			}
-		]);
-		expect(sceneData.combatants.find((c) => c.id === 'GandalfId')?.conditionStates).toEqual([
-			{
-				activeLevel: 3,
-				id: 'burning',
-				activeSinceTick: 4
-			}
-		]);
-
-		// Toggle burning away from Gandalf
-		toggleCondition('GandalfId', 'burning');
-		expect(sceneData.combatants.find((c) => c.id === 'GandalfId')?.conditionStates).toEqual([]);
-
-		// When burning is readded it starts at Gandalfs current offset
-		toggleCondition('GandalfId', 'burning');
-		expect(sceneData.combatants.find((c) => c.id === 'GandalfId')?.conditionStates).toEqual([
-			{
-				activeLevel: 1,
-				id: 'burning',
-				activeSinceTick: 9
-			}
-		]);
-	});
 });
 
 describe('set combat states', () => {
